Add basic OAuth2 implementation
Basically created a users table in the database. The user is able to
register using the /registration endpoint and login using the /token
endpoint. The first endpoint secured is POST /wordlists. I still need to
update the schema so that users are associated with word lists. I also
need to create the registration and login UI.